design prompts added
If the system detects dizziness, and symptoms of fainting, it will give a prompt on the connected devices to ensure whether the user is safe or not. If yes, the system is sorted but if there is no response (or NO in that case), immediately the vehicle is stopped safely using mechanics and the concerned authorities (hospital, police) are informed via speed-dial or satellite communications.